+++ packages/clerk-js/src/ui/contexts/components/UserProfile.ts
@@ -9,6 +9,7 @@ import type { ParsedQueryString } from '../../router';
 import { useRouter } from '../../router';
 import type { UserProfileCtx } from '../../types';
 import { useEnvironment } from '../EnvironmentContext';
+import { useStatements, useSubscription } from './Plans';
 
 type PagesType = {
   routes: NavbarRoute[];
@@ -21,6 +22,7 @@ export type UserProfileContextType = UserProfileCtx & {
   authQueryString: string | null;
   pages: PagesType;
   shouldAllowIdentificationCreation: boolean;
+  shouldShowBilling: boolean;
 };
 
 export const UserProfileContext = createContext<UserProfileCtx | null>(null);
@@ -32,6 +34,20 @@ export const useUserProfileContext = (): UserProfileContextType => {
   const environment = useEnvironment();
   const { user } = useUser();
 
+  const subscription = useSubscription();
+  const statements = useStatements();
+
+  const hasNonFreeSubscription = subscription.data?.subscriptionItems.some(item => item.plan.hasBaseFee);
+
+  // TODO(@BILLING): Remove this when C1s can disable user billing seperately from the organization billing.
+  const shouldShowBilling =
+    // The instance has at lease one visible plan the C2 can choose
+    environment.commerceSettings.billing.user.hasPaidPlans ||
+    // The C2 has a subscription, it can be active or past due, or scheduled for cancellation.
+    hasNonFreeSubscription ||
+    // The C2 had a subscription in the past
+    Boolean(statements.data.length > 0);
+
   if (!context || context.componentName !== 'UserProfile') {
     throw new Error('Clerk: useUserProfileContext called outside of the mounted UserProfile component.');
   }
@@ -39,8 +55,8 @@ export const useUserProfileContext = (): UserProfileContextType => {
   const { componentName, customPages, ...ctx } = context;
 
   const pages = useMemo(() => {
-    return createUserProfileCustomPages(customPages || [], clerk, environment);
-  }, [customPages]);
+    return createUserProfileCustomPages(customPages || [], clerk, shouldShowBilling, environment);
+  }, [customPages, shouldShowBilling]);
 
   const shouldAllowIdentificationCreation = useMemo(() => {
     const { enterpriseSSO } = environment.userSettings;
@@ -62,5 +78,6 @@ export const useUserProfileContext = (): UserProfileContextType => {
     queryParams,
     authQueryString: '',
     shouldAllowIdentificationCreation,
+    shouldShowBilling,
   };
 };