+++ packages/chrome-extension/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+## 2.5.29
+
+### Patch Changes
+
+- Updated dependencies [[`73fe6ff`](https://github.com/clerk/javascript/commit/73fe6ff4be665b5543b865d303dc51060c9f00bb), [`9796fbf`](https://github.com/clerk/javascript/commit/9796fbf57494e108bdb531accc26bdb8e54e59f9), [`375ad8d`](https://github.com/clerk/javascript/commit/375ad8dd8668b8bacc048aaf8135b5d1c62e00ea), [`f28179b`](https://github.com/clerk/javascript/commit/f28179b30102550cc5601fcd22da7ca9a90959ee), [`e6e19d2`](https://github.com/clerk/javascript/commit/e6e19d2d2f3b2c4617b25f53830216a1d550e616), [`e6e19d2`](https://github.com/clerk/javascript/commit/e6e19d2d2f3b2c4617b25f53830216a1d550e616), [`1b1e8b1`](https://github.com/clerk/javascript/commit/1b1e8b1fd33b787f956b17b193e5fd0a4cdc6cec)]:
+  - @clerk/clerk-js@5.91.2
+  - @clerk/clerk-react@5.46.1
+  - @clerk/shared@3.24.1
+
 ## 2.5.28
 
 ### Patch Changes